requirements.txt
torch>=2.0.0
transformers>=4.30.0
python-telegram-bot==13.15
huggingface-hub>=0.16.0
safetensors
